Combined fat-cored and colliding vortex rings in one file
