+++ amr-wind/physics/CMakeLists.txt
@@ -12,7 +12,7 @@ target_sources(${amr_wind_lib_name}
   EkmanSpiral.cpp
   SyntheticTurbulence.cpp
   HybridRANSLESABL.cpp
-  FatCoredVortexRing.cpp
+  VortexRing.cpp
   )
 
 add_subdirectory(multiphase)
